neural_extra/requirements.txt
numpy
imageio
Pillow  # provides PIL
scipy==1.1.0
tensorflow-gpu >= 1.0,<2.0  # installs tensorflow with GPU support, should still work even without GPU
